+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+- `Textarea` Added an option to pass `onInvalidCapture` as an attribute. 
 - `TextField` Added support for colored description based on the state of the TextField
 
 ### Fixed